\name{dM1}
\alias{dM1}
\title{
Distance measure defined in Eq. (3.1) of Li et al. (2016)
}
\description{
Compute the distance measure defined in Eq. (3.1) of Li et al. (2016) to compare the factor loading matrix in its Monte Carlos experiments.
}
\usage{
dM1(A, Ahat)
}
\arguments{
  \item{A}{
The original factor loading matrix A
}
  \item{Ahat}{
The estimated factor loading matrix A}
}

\references{
Li, W., Gao, J., Li, K., & Yao, Q. (2016). Modeling Multivariate Volatilities via Latent Common Factors. Journal of Business & Economic Statistics, 34(4), 564-573.
}
\author{
Carlos Trucios
}
\keyword{Factor loading matrix}
\keyword{DM1}
